+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -10,15 +10,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added a new API to bake HDRP probes from C# (case 1276360)
 - Added support for pre-exposure for planar reflections.
 - Added support for nested volume components to volume system.
+- Added a toggle to allow to include or exclude smooth surfaces from ray traced reflection denoising.
 
 ### Fixed
 - Fixed probe volumes debug views.
+- Fixed couple samplers that had the wrong name.
 
 ### Changed
 - Removed the material pass probe volumes evaluation mode.
 - Volume parameter of type Cubemap can now accept Cubemap render textures and custom render textures.
 - Removed the superior clamping value for the recursive rendering max ray length. 
 - Removed the superior clamping value for the ray tracing light cluster size.
+- When in half res, RTR denoising is executed at half resolution and the upscale happens at the end.
+- Removed the upscale radius from the RTR.
 
 ## [10.3.0] - 2020-12-01